Modify index files in models for testing sequel and pg
