+++ svbackend/models/index.js
@@ -1,20 +1,43 @@
-const sequelize = require('../utilities/dbPool');
-const Admin = require('./Admin');
-const AdminRole = require('./AdminRole');
-const User = require('./User');
-const Subscription = require('./Subscription');
-const Student = require('./Student');
-const Course = require('./Course');
-const Module = require('./Module');
-const Lesson = require('./Lesson');
-const Project = require('./Project');
-const Grade = require('./Grade');
+'use strict';
 
-(async () => {
-    try {
-        await sequelize.sync({ alter : true });
-        console.log("Database & tables created!");
-    } catch (error) {
-        console.error("Error creating tables:", error);
-    }
-})();
+const fs = require('fs');
+const path = require('path');
+const Sequelize = require('sequelize');
+const process = require('process');
+const basename = path.basename(__filename);
+const env = process.env.NODE_ENV || 'development';
+const config = require(__dirname + '/../config/config')[env];
+const db = {};
+
+let sequelize;
+if (config.use_env_variable) {
+  sequelize = new Sequelize(process.env[config.use_env_variable], config);
+} else {
+  sequelize = new Sequelize(config.database, config.username, config.password, config);
+}
+
+fs
+  .readdirSync(__dirname)
+  .filter(file => {
+    return (
+      file.indexOf('.') !== 0 &&      // Ignore hidden files
+      file !== basename &&            // Ignore this index.js file itself
+      file.slice(-3) === '.js' &&     // Only include .js files
+      file.indexOf('.test.js') === -1 // Ignore test files
+    );
+  })
+  .forEach(file => {
+    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
+    db[model.name] = model;
+  });
+
+Object.keys(db).forEach(modelName => {
+  if (db[modelName].associate) {
+    db[modelName].associate(db);
+  }
+});
+
+db.sequelize = sequelize;
+db.Sequelize = Sequelize;
+
+module.exports = db;
